Resolve issue #1302:
Handle the case where a CVS file is first added on a branch, so the
RCS file starts out in the Attic with revision 1.1 in state "dead",
and the case where two successive revisions are both in state "dead".

These two cases actually overlap, and can probably be handled by
shared logic.  But that will require some other changes, and my first
priority is to get these bugfixes in for the 0.23 release.

* tools/cvs2svn/cvs2svn.py
  (TreeMirror.delete_path): If the path can't be reached, return None. 
  (Dump.delete_path): If the path didn't exist in the head tree
    mirror, then don't output a delete to the dumpfile.  Return 1 if
    actually deleted a path, None if not.
  (Commit.commit): Don't even bother to delete revision 1.1 of a file,
    and don't output a revision unless there was some activity.


git-svn-id: http://cvs2svn.tigris.org/svn/cvs2svn/trunk@482 be7e6eca-30d4-0310-a8e5-ac0d63af7087