Tests and helper functions should not print to stdout.
Helper functions should not print directly to stdout, because then it
is impossible to suppress the output for errors that are expected.
Instead, they should store the error message in the svntest.Failure
instance that they raise.  Then: (1) if the exception is caught (i.e.,
because an error is expected), the error message is not displayed; (2)
if the exception is not caught, the error message is printed out by
svntest.TestCase.run().

* run-tests.py (check_rev, ensure_conversion, show_usage,
  prune_with_care, ctrl_char_in_log, check_props): Store error
  messages into svntest.Failure instance instead of printing them
  directly to stdout.


git-svn-id: http://cvs2svn.tigris.org/svn/cvs2svn/trunk@1485 be7e6eca-30d4-0310-a8e5-ac0d63af7087