Removed level of indirection in SymbolicNameFillingGuide data structure.
SymbolicNameFillingGuide used to store all nodes in one giant
two-level dictionary self._node_tree in the form

    { key : { pathname_component : key } }

, where 'key' is a synthetically-generated unique key.  Going one level
into the directory tree requires two dictionary lookups: one to find the
key of the subdirectory, then another to find the node describing the
subdirectory.

This commit removes the indirection by storing the information as a
hierarchy of nested dictionaries of the form

    { pathname_component : subnode }

, where each subnode is a dictionary of the same form.

I would expect a slight improvement in runtime and memory space,
though no significant difference is seen when executing the test
suite.

* cvs2svn (SymbolicNameFillingGuide): Changed storage scheme for
  _node_tree, as described above.

  (SymbolicNameFillingGuide._get_node_for_path): Renamed from
  _get_node_key(), to reflect its changed meaning.

  (SymbolicNameFillingGuide.get_node): Removed (no longer needed).

  (SymbolicNameFillingGuide.get_best_revnum): Interpret "node"
  argument as a node (previously it was interpreted as a node_key,
  despite its name).

  (SymbolicNameFillingGuide._list_revnums): Renamed from
  _list_revnums_for_key(), to reflect its changed meaning.

  (SymbolicNameFillingGuide.print_node_tree): Changed format of output
  to match new data layout.

  (FillSource.__init__, SVNRepositoryMirror._fill): Changed callers.


git-svn-id: http://cvs2svn.tigris.org/svn/cvs2svn/trunk@1623 be7e6eca-30d4-0310-a8e5-ac0d63af7087