Prepare to use a new object as the cvs2svn_rcsparse.Sink in pass1.
The CollectData class is used both to collect global information about
the CVS repository (e.g., database connections, fatal_errors,
num_files, etc.) and also as the cvs2svn_recparse.Sink to receive
information about individual files (e.g., branch_names, file_in_attic,
etc.).  This is currently accomplished by initializing a bunch of
file-specific variables in the set_fname() method that are used in the
parsing of one file then forgotten.

I would like to rewrite this so that a new FileDataCollector class is
responsible for collecting the data for a particular file, and
CollectData creates instances of the new class as needed.  The new
class will have a reference to the CollectData object so that it can
write its information into the global databases.

* cvs2svn (CollectData.get_file_data_collector): New method.

  (CollectData._set_fname): Method renamed from set_fname().

  (pass1.visit_file): Use get_file_data_collector() method.


git-svn-id: http://cvs2svn.tigris.org/svn/cvs2svn/trunk@1806 be7e6eca-30d4-0310-a8e5-ac0d63af7087