Record how many times each branch serves as a possible parent of a symbol.
This is work towards better choice of symbol sources, imitating the
algorithm developed by Daniel Jacobowitz <drow@false.org>.

For now, we only record the obvious possible parents (not the ones
having to do with non-trunk default branches).

The recording of "possible parents" is surprisingly expensive in terms
of CPU time on my gcc test repository, because the repository has lots
of symbols, many with 100+ possible parents and the algorithm is
O(N^2) in the number of possible parents for a given symbol.
Hopefully this is atypical.

* cvs2svn_lib/symbol_statistics.py (_Stats.possible_parents): New
  member.

  (_Stats.__init__): Initialize new member.

  (_Stats.register_possible_parent): Allow new member to be updated.

  (_Stats.__repr__): New method.  Include str(self) plus statistics
  for all of the possible parents.

* cvs2svn_lib/collect_data.py
  (_FileDataCollector._register_branch_possible_parents,
  _FileDataCollector._register_tag_possible_parents): New methods.

  (_FileDataCollector.tree_completed): Invoke new methods.

* doc/symbol-notes.txt: Add note that CollectRevsPass stores
  possible_parent statistics.  Start adding docs for OutputPass.


git-svn-id: http://cvs2svn.tigris.org/svn/cvs2svn/trunk@3055 be7e6eca-30d4-0310-a8e5-ac0d63af7087