Make SymbolStatistics.check_consistency() sanity checks more robust.
We want to verify that the SYMBOLS argument unambiguously defines how
each symbol should be converted (or excluded).  Previously, this was
tested only by verifying that the length of the SYMBOLS argument was
equal to the number of symbols in SymbolStatistics.  Now it checks
that there is a 1:1 correspondence between the two.

These checks should only fail if the SymbolStrategy machinery is not
working right, *not* in the case of a user-level failure to specify
rules for all symbols.  (Unless of course the user writes his own
buggy SybmolStrategy class. :-) )

* cvs2svn_lib/symbol_statistics.py
  (SymbolStatistics.check_consistency): Make sanity checks more
  robust.  Raise InternalError in the case of problems.


git-svn-id: http://cvs2svn.tigris.org/svn/cvs2svn/trunk@3492 be7e6eca-30d4-0310-a8e5-ac0d63af7087