Sort branches in apparent creation order rather than numerical order.
If a branch is created then deleted, a later branch can be assigned
the recycled branch number; therefore branch numbers don't have much
to do with branch creation order.  But it appears that symbols are
listed in the header of the RCS file in the reverse order that they
were created.  So sort them that way in rev_data.branches_data.


* cvs2svn_lib/collect_data.py (_FileDataCollector._sort_branches):
  Sort branches in the reverse order that they were listed in the RCS
  file symbol section rather than in numerical order.


git-svn-id: http://cvs2svn.tigris.org/svn/cvs2svn/trunk@3549 be7e6eca-30d4-0310-a8e5-ac0d63af7087