Prevent a stack overflow by preventing recursion when discard() is called.
If a long chain of revisions is deleted, then each deletion can
trigger the next one to be deleted, resulting in stack growth
proportional to the number of revisions in the chain.  This change
prevents that (at least in the case of discard()) by deferring
deletions while an outer discard() is running.

* cvs2svn_lib/checkout_internal.py
  (TextRecordDatabase.deferred_deletes): New member.

  (TextRecordDatabase.__init__, TextRecordDatabase.__setstate__):
  Initialize new member.

  (TextRecordDatabase.discard): Prevent recursion by deferring calls
  to discard() that don't occur at the outermost level.  Also, allow
  multiple arguments.

  (TextRecordDatabase.free_unused): Pass all of the unused text record
  ids to one call of discard().


git-svn-id: http://cvs2svn.tigris.org/svn/cvs2svn/trunk@3684 be7e6eca-30d4-0310-a8e5-ac0d63af7087