Omit SymbolChangesets that have no non-Noop CVSSymbols.
Please note that this changes the observed behavior of the program.
If a symbol is based on a dead revision, then previously an empty
directory would be created for the symbol even if no files were ever
stored there.  Now a directory is only created when and if a non-dead
file is added to the symbol.

This change clears up most "empty SymbolFillingGuide" warnings.

* cvs2svn_lib/svn_commit_creator.py
  (SVNCommitCreator._process_tag_changeset,
  SVNCommitCreator._process_branch_changeset): Filter out
  CVSSymbolNoops.  If no CVSSymbols are left, don't generate an
  SVNCommit.

* run-tests.py (PhoenixBranch.run, no_spurious_svn_commits,
  branch_from_default_branch): Adjust results to match new expected
  behavior.

  (test_list): Remove the XFail annotation from
  no_spurious_svn_commits() and branch_from_default_branch().


git-svn-id: http://cvs2svn.tigris.org/svn/cvs2svn/trunk@3730 be7e6eca-30d4-0310-a8e5-ac0d63af7087