Omit the initial 'dead' branch revisions when a file is added on a branch.
Newer versions of CVS add an initial 'dead' revision on a branch when
a file is added on a branch.  (This is to record the fact that the
file was not present between the time that the branch was created and
the time when it was actually added.)  With this change, cvs2svn omits
such commits and severs the first true branch revision from the source
branch so that it is treated as a pure add.

* cvs2svn_lib/cvs_file_items.py
  (CVSFileItems._initial_branch_delete_unneeded,
  CVSFileItems.remove_initial_branch_deletes): New methods.

* cvs2svn_lib/collect_data.py (_ProjectDataCollector._process_file):
  Call CVSFileItems.remove_initial_branch_deletes().

* run-tests.py (double_branch_delete): Adjust test to reflect new
  expected results.

  (add_on_branch): Check that the expected sequence of commits occurs.

* CHANGES: Document change to behavior.


git-svn-id: http://cvs2svn.tigris.org/svn/cvs2svn/trunk@3773 be7e6eca-30d4-0310-a8e5-ac0d63af7087