Compute the relative order of CVSPaths and record it in the instances.
CVSPath.ordinal is an integer that is used as the first sort criterion
for CVSPath objects; if the ordinals are the same then the cvs_path
members are compared.  (Please note that CVSPath.cvs_path is a
computed property and its computation is somewhat expensive.)

When CVSPaths are first constructed, the ordinal instance is
initialized to 0 for CVSDirectories and 1 for CVSFiles.  This makes
CVSDirectories sort before CVSFiles, with alphabetic sorting by
cvs_path within each class.

Before the CVSPaths are pickled to the file, they are sorted (using
the ordering just described) then the ordinal members are set to the
index of the CVSPath in the fully-ordered list.  Thereafter, the
ordinal members are all unique, so the comparison is always determined
by the ordinal member.  This avoids the expense of computing the
cvs_path for each comparison.

* cvs2svn_lib/cvs_file.py (CVSPath.ordinal): New member.

  (CVSPath.__slots__): Add new member to slot list.

  (CVSPath.__init__, CVSPath.__getstate__, CVSPath.__setstate__):
  Initialize, save, and restore new member.

  (CVSPath.__cmp__): Use the ordinal member as the default comparison
  criterion.

  (CVSDirectory.__cmp__, CVSFile.__cmp__): Remove methods (the
  inherited version is fine now).

* cvs2svn_lib/cvs_file_database.py (CVSFileDatabase.close): Before
  writing out the file list, sort the files and store their ordinal
  in each instance.


git-svn-id: http://cvs2svn.tigris.org/svn/cvs2svn/trunk@4095 be7e6eca-30d4-0310-a8e5-ac0d63af7087