If a path has not been copied, then leave parent_source unset.
Previously, the parent_source argument was passed to _fill_directory()
regardless of whether the parent was actually copied.  This imposed an
unnecessary constraint on how subdirectories are filled (essentially
adding a bias to filling sibling subdirectories from the same
source).  This change removes that bias.

In practice, this change should hardly make any difference.

* cvs2svn_lib/svn_repository_mirror.py
  (SVNRepositoryMirror._fill_directory): Only pass a new parent_source
  for subdirectories if the directory itself was copied.


git-svn-id: http://cvs2svn.tigris.org/svn/cvs2svn/trunk@4119 be7e6eca-30d4-0310-a8e5-ac0d63af7087