Merge the repository-mirror branch (revisions 4432:HEAD) to trunk.
The only user-visible change is that cvs2git avoids creating a fixup
branch if the tag contents are identical to those of its parent.

Under the covers, this change involves abstracting out a
RepositoryMirror class that can keep track of which files and
directories are present in every version of every line of development
in a form that can also be used by the Git output code.  (Previously
this functionality was strongly intertwined with the SVN output code.)
This is not simple because this code is a performance bottleneck, so a
certain intimacy is required between it and its users to avoid
excessive lookups.

For SVN output, the new code is still slower than the code that it
replaced (it increases the time for OutputPass by about 10%) and the
RepositoryMirror interface is rather complicated--intuitive and
straightforward for 95% of its uses, but there are a couple of
operations that require extra care not to screw up (e.g., see
repository-mirror r4473 and r4490).

There is still scope for simplification and optimization, but I'll
continue that work on trunk.

Files changed:

* CHANGES
* run-tests.py
* doc/design-notes.txt
* cvs2svn_lib/config.py
* cvs2svn_lib/cvs_file.py
* cvs2svn_lib/dumpfile_delegate.py
* cvs2svn_lib/fill_source.py
* cvs2svn_lib/git_output_option.py
* cvs2svn_lib/repository_mirror.py
* cvs2svn_lib/stdout_delegate.py
* cvs2svn_lib/svn_output_option.py
* cvs2svn_lib/svn_repository_delegate.py
* cvs2svn_lib/svn_repository_mirror.py
* cvs2svn_lib/svn_revision_range.py
* test-data/mirror-keyerror2-cvsrepos/*
* test-data/mirror-keyerror3-cvsrepos/*


git-svn-id: http://cvs2svn.tigris.org/svn/cvs2svn/trunk@4518 be7e6eca-30d4-0310-a8e5-ac0d63af7087