Under Windows, invoke sort from PATH rather than the Windows sort.
Patch by: James Abbatiello <abbeyj@gmail.com>

James's explanation:

Trying to call out to GNU sort always uses Windows sort even if GNU
sort is available and appears in the PATH before Windows sort.  The
subprocess module uses the Win32 CreateProcess which always looks in
the Windows system32 directory before it looks in the directories
listed in the PATH environment variable.  Since the Windows sort.exe
is in the system32 directory it will always be chosen.  A simple
workaround is to launch the sort in a shell.  When the shell (cmd.exe)
searches it only examines the directories in the PATH so putting the
directory with GNU sort ahead of the Windows system32 directory will
cause GNU sort to be chosen.

* cvs2svn_lib/passes.py (sort_file): Under Windows, invoke the sort
  command with shell=True, which causes the sort command to be sought
  in PATH before in the system32 directory.

git-svn-id: http://cvs2svn.tigris.org/svn/cvs2svn/trunk@4820 be7e6eca-30d4-0310-a8e5-ac0d63af7087