Create test-case related paths using hashs instead of long strings.
Patch by: James Abbatiello <abbeyj@gmail.com>
Modified by me.

James's explanation:

The symbol_transform test creates a directory with a very long name.
Some files inside this directory eventually exceed the Windows limit
of 256 characters in an absolute path which causes the test to fail.
I've tried to work around this by using a hash of the arguments in the
directory name instead of using them literally.  This keeps the length
down and preserves the existing guarantee that directory names won't
have illegal characters in them.  The downside is that the directory
names in the temp directory become somewhat harder to decipher.

* run-tests.py (make_conversion_id): Create the conversion id based on
  a hash of (most of) the conversion parameters, to prevent the ids
  from becoming too long.

git-svn-id: http://cvs2svn.tigris.org/svn/cvs2svn/trunk@4825 be7e6eca-30d4-0310-a8e5-ac0d63af7087