Improve FAQ about building Python on Mac OS X.
Patch by: Vincent McIntyre <Vince.McIntyre@atnf.csiro.au>

From his email (lightly edited):

The instructions for building cvs2svn on Mac OS X seem to be
incorrect.  The reason is that the python2.6 configure options
--enable-shared and --enable-framework are mutually exclusive. See
http://bugs.python.org/issue5809.

I tested the revised procedure on 10.5.7 (intel) and it worked fine.
I built gdbm-1.8.3 and Python-2.6.2, with the system compiler
(gcc-4.0.1).

If I try to build the way currently stated in the FAQ, I get the same
error as shown in #5809 above.

* www/faq.html: Improve FAQ about building Python on Mac OS X.

git-svn-id: http://cvs2svn.tigris.org/svn/cvs2svn/trunk@4847 be7e6eca-30d4-0310-a8e5-ac0d63af7087