Loosen criteria for deleting initial trunk commits.
Do not require the filename to match in the log message of an initial
trunk delete to declare it unneeded.  This criterion did not really
add to the safety of the transformation and it also caused the
transformation to be skipped if the RCS file had ever been renamed.

Also do not check the branch named in the log message, since that
branch might not be the first in the list of branches.

* cvs2svn_lib/cvs_file_items.py
  (CVSFileItems._is_unneeded_initial_trunk_delete): Do not require
  filename or branch name match in log message.

git-svn-id: http://cvs2svn.tigris.org/svn/cvs2svn/trunk@4920 be7e6eca-30d4-0310-a8e5-ac0d63af7087