Fail if the PipeStream command outputs anything to stderr.
In PipeStream.close(), check not only the exit status of the program
but also whether it output anything to standard error.  In either
case, consider the program to have failed and emit a CommandError
exception.

This should hopefully catch some other possible failures of the "cvs"
or "co" commands.

It is possible that some old or deviant version of these programs
emits output on stderr even though the command was successful.  If
this turns out to be the case, then this change will have to be
revisited.

git-svn-id: http://cvs2svn.tigris.org/svn/cvs2svn/trunk@4961 be7e6eca-30d4-0310-a8e5-ac0d63af7087