Add an --include-empty-directories option.
When this option is set, treat an empty directory within the CVS
repository as a part of the source tree, and create it in the SVN
repository at the point when the directory's parent is created.  (An
empty directory that is directly in a project root directory is
created when the project is initialized.)

Show, in cvs2svn-example.options, how to accomplish the equivalent of
--include-empty-directories via an options file.

Add tests of this new feature.

This commit fixes issue #31.

git-svn-id: http://cvs2svn.tigris.org/svn/cvs2svn/trunk@5027 be7e6eca-30d4-0310-a8e5-ac0d63af7087