Handle --default-eol=binary as per documentation.
Setting --default-eol=binary has been broken since r4687.  The error
has presumably gone undetected because the default value was treated
as "binary" (so --default-eof=binary never needed to be used
explicitly) but handled internally as None (which wasn't broken).

The breakage resulted in a crash, so there is no risk of hidden data
corruption.

git-svn-id: http://cvs2svn.tigris.org/svn/cvs2svn/trunk@5293 be7e6eca-30d4-0310-a8e5-ac0d63af7087