cvs2git: clarify documentation of --use-external-blob-generator
The blob generator runs without heavy supervision by the main cvs2git
script, which is great for parallelism.  The only gotcha is that the
blob generator and main cvs2git need access to separate parts of the
fast-import stream, which breaks the pattern

	cvs2git <options> | git fast-import

that was great for parallelism.  Document this requirement (blobs
and revisions must go to different files) to avoid confusion.

Can we salvage the parallelism by streaming blobs directly to
fast-import and queuing up revs in a file?  Alas, no: the blob
generator seeks around in its output file to be able to read back old
blobs.  Document that requirement, too (the blob file cannot be a
pipe).

Balance out the new cautionary notes by removing an old one: the
--use-external-blob-generator code is solid and no longer deserves
the fearsome "experimental" tag.

Signed-off-by: Jonathan Nieder <jrnieder@gmail.com>

git-svn-id: http://cvs2svn.tigris.org/svn/cvs2svn/trunk@5304 be7e6eca-30d4-0310-a8e5-ac0d63af7087