Improve error reporting of ValueExceptions emitted by rcsparse.
Patch by: Craig Leres <leres@tigris.org>
(Submitted with issue #141.)

Comment (copied from issue):

I had a really old RCS ,v file that had a bad year in a date string
(100.01.05.12.00.52) and it was a fair amount of work to figure out
what the problem was.

In the process I made a couple of short patches to cvs2svn to provide
additional info when the error is "invalid year"; patches attached.

git-svn-id: http://cvs2svn.tigris.org/svn/cvs2svn/trunk@5308 be7e6eca-30d4-0310-a8e5-ac0d63af7087