verify-cvs2svn.py: Do not blithely discard the last character from each line.
SvnRepos.list() was discarding the last character of each line of
output from "svn ls $URL", under the assumption that the last character
would be a "/".  But that assumption is incorrect if the directory
contains a file (as opposed to only subdirectories).  So change the
code to use rstrip() to remove trailing slash characters.

git-svn-id: http://cvs2svn.tigris.org/svn/cvs2svn/trunk@5366 be7e6eca-30d4-0310-a8e5-ac0d63af7087