ExternalBlobGenerator: Skip over files that have no marks.
It is possible that *no* revisions will be needed from an RCS file
(see previous commit message for an example).

In such a case, don't pass the file along to
external_blob_generator.py, because it doesn't do anything useful
anyway.

This is also a belt-and-suspenders second fix to the same bug fixed by
the previous commit.

git-svn-id: http://cvs2svn.tigris.org/svn/cvs2svn/trunk@5406 be7e6eca-30d4-0310-a8e5-ac0d63af7087