Write all progress information to stderr rather than stdout.
It used to be that errors were written to stderr and other progress
information was written to stdout.  Change to writing *all* progress
information on stderr, to leave open the possibility of writing data
(e.g., git dumpfiles) to stdout to allow cvs2svn to be used in a
pipeline.

The main difficulty is that the test suite used to assume that any
output written to stderr indicated a failure of the conversion, so the
increased output on stderr would be interpreted as failures.  To get
the same effect, change the tests to run cvs2svn with the verbosity
turned down all the way so that only errors are written to stderr (as
before).

Finally, there were a couple of tests that looked for warnings on the
command's stdout.  Change these tests to increase the command
verbosity and look for the same messages on stderr.

git-svn-id: http://cvs2svn.tigris.org/svn/cvs2svn/trunk@5419 be7e6eca-30d4-0310-a8e5-ac0d63af7087