cvs2git: Emit some more information in OutputPass.
During OutputPass of cvs2git, output the correspondence between the
SVN revnums used in CreateRevsPass and the commit marks written to the
dumpfile.  This gives users a hope of determining which CVS file
revisions are added in which Git commits (though they still have to
convert marks -> SHA1s themselves).

git-svn-id: http://cvs2svn.tigris.org/svn/cvs2svn/trunk@5432 be7e6eca-30d4-0310-a8e5-ac0d63af7087