If SOURCE_DATE_EPOCH is set, use it for manpages
To improve build reproducibility, use the value of the environment
variable SOURCE_DATE_EPOCH as the date to embed in manpages if it is
set, rather than unconditionally using the current date. The idea is
that packaging systems can feed in the date of the last source
modification, so that the result is the same no matter when the package
is built. See the SOURCE_DATE_EPOCH specification [1] for more
information.

This was reported as Debian bug #831066 [2].

[1] https://reproducible-builds.org/specs/source-date-epoch/
[2] https://bugs.debian.org/831066

Suggested-by: Chris Lamb <lamby@debian.org>

git-svn-id: http://cvs2svn.tigris.org/svn/cvs2svn/trunk@5465 be7e6eca-30d4-0310-a8e5-ac0d63af7087